backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.0
httpx==0.27.0
pydantic==2.5.0
sqlalchemy==2.0.35
python-multipart==0.0.9
reportlab==4.4.10
